Move typeorm module config closer to the service that uses it
