+++ application/services/telegraph/src/storage/storage.module.ts
@@ -1,10 +1,10 @@
 import {Module} from '@nestjs/common'
 import {TypeOrmModule} from '@nestjs/typeorm'
 
-import {getTypeormModuleOptions} from '@/storage/common/typeorm-module-options'
 import {UserEntity} from '@/storage/entities/user.entity'
 import {UserRepository} from '@/storage/repositories/user/user.repository'
 import {StorageConfigService} from '@/storage/services/storage-config.service'
+import {getTypeormModuleOptions} from '@/storage/services/typeorm-module-options'
 
 @Module({
   exports: [UserRepository, StorageConfigService],
